nxp-s32g2xx: update the u-boot version to 2019.04
Issue: LINCD-2894

The NXP SDK bsp25.0 should use the u-boot version with 2019.04.
But since there is a u-boot bug with "tftp hangs when ring buffer becomes full".
So keep the u-boot version with 2018.07, now the NXP SDK gives a draft
fix for this issue. So update the u-boot version to 2019.04.

Signed-off-by: Xulin Sun <xulin.sun@windriver.com>