nxp-s32g2xx: clear COMPATIBLE_MACHINE variable
Issue: LINCD-3696

This defect reports some pfe packages are not shipped if building qemumips,
Actually the issue is not reproduced on my local environment even
builing qemumips or nxp-s32g2xx.

These reported missing packages "pfeng.ko" are as a part of deploy-rpm
images, which will dynamic installed to the rootfs.

Anyway it should not build the pfe package on the outside of this bsp
layer. So clear the COMPATIBLE_MACHINE first to set with "nxp-s32g2xx"
to avoid it confusion.

Signed-off-by: Xulin Sun <xulin.sun@windriver.com>