nxp-s32g: [V5] update uboot, pfe, kernel to bsp28 pre-release version
Issue: LINUXEXEC-12313

All the U-Boot, kernel and pfe are updated to bsp28 version.
And pfe version is 0.9.3.

In V2 version, the gpio issue has been fixed. And ATF feature
is added.

In V3 version, fix pfe build warning and debug calltrace.

In V4 version, add the latest uboot, atf, kernel patches and
add support for pkcs11-hse.

In V5 version, modify the atf bb file to fetch code from local
package.

NOTICE:
1. from V3 version, please use git am --keep-cr to apply this
patch, or the pfe build will fail due to do_patch(). The root
cause is that the there is a patch for pfe_emac.c which is edited
with Windows OS, the end of each line will have a ^M charachter,
we should keep the ^M when git am to make the do_patch() success.

2. add WRLINUX_SUPPORTED_RECIPE_pn-libp11 = '2' into
<build-path>/conf/local.conf for fix libp11 building warning.

3. add WRLINUX_SUPPORTED_RECIPE_pn-arm-trusted-firmware = '2' into
<build-path>/conf/local.conf for fix libp11 building warning.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>