nxp-s32g2xx: add ATF(arm-trusted-firmware) support
Issue: LINUXEXEC-12756

The ATF feature is power management and trusted environment related,
for S32G platform now, it is mainly for power management like STR(Sleep
To Ram or S3 sleep) or CPU hotplug. By default, this feature is not
built in with WIC image, if users want to install it, please add the
following line in <build-dir>/conf/local.conf of project:
	IMAGE_INSTALL += "arm-trusted-firmware"

Then the ATF image will be integrated into the WIC image automatically.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>