nxp-s32g2xx: add pkcs11-hse support
Issue: LINUXEXEC-12756

Previously, there is no demo for pkcs11 and hse on linux system
for S32G platform, now the NXP SDK bsp28 has been released, and
the demo for pkcs11 & hse is added for linux. However, the related
git repo for this demo is unavailable without NXP NDA, so the source
code is in a .tar.gz package with BSD license. Further more, in demo
source code, there have a wrong parameter, so add one patch to fix it
to make the demo runs OK.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>