nxp-s32g2xx: add IPCF(Inter-Platform Communication Framework) support
Issue: LINUXEXEC-12756

There are two types of CPU cores on S32G platform, A53 and M7.
The IPCF is mainly for communication between them. This patch
is to add support for this feature.

Now this feature is not enabled by default, if users want to enable
it, please add the following line in <build-dir>/conf/local.conf:
	PNWHITELIST_wr-nxp-s32g2xx += "ipc-shm"
	IMAGE_INSTALL += "ipc-shm"

then the feature will be integrated into WIC image.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>