pfe: Fix for installed-vs-shipped error
Issue: LINCD-4747

Fixed:
WRTEMPLATE += "feature/ostree"
$ MACHINE=nxp-s32g2xx bitbake pfe
ERROR: pfe-0.9.3-r0 do_package: QA Issue: pfe: Files/directories were installed but not shipped in any package:
  [snip]
  /lib/modules/5.4.105-yocto-standard/kernel/drivers/net/ethernet/nxp/pfe/pfeng.ko

This error happens when build with ostree since it enables usrmerge. The
pfeng.ko is packed by base_libdir which is '/lib' when no usrmerge, but it would
be /usr/lib when usrmerge is enabled, then pfeng.ko is not packed, which causes
the error.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>