nxp-s32g2xx: drop hse support on RDB2
Issue: LINCD-4910

Previously, we added hse support on RDB2, but it affacts the
functionality of ATF image. And for hse feature on RDB2, it still
needs equipment provider to give a proper solution in the end,
so drops hse feature on RDB2 temporarily. And we will add it back
when this issue is totally fixed.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>