nxp-s32g2xx.conf: Add virtual/bootloader to MULTI_PROVIDER_WHITELIST
Issue: LINCD-4992

Fixed when multilib is enabled:
$ bitbake u-boot-s32

ERROR: Nothing PROVIDES 'u-boot-s32'
u-boot-s32 was skipped: We shouldn't have multilib variants for u-boot-s32
u-boot-s32 was skipped: PREFERRED_PROVIDER_virtual/lib32-bootloader set to lib32-u-boot-s32, not u-boot-s32

This is because PREFERRED_PROVIDER_virtual/bootloader is set to u-boot-s32, and
it would be extended to:
PREFERRED_PROVIDER_virtual/lib32-bootloader = "lib32-u-boot-s32"

But u-boot-s32 is in NON_MULTILIB_RECIPES, so there is no lib32-u-boot-s32, and
the error happens. The PREFERRED_PROVIDER_virtual/bootloader must be set to
make sure u-boot-s32 is built, otherwise, it may fail to build with ostree, so
add it to MULTI_PROVIDER_WHITELIST to fix the error.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>