nxp-s32g2xx: add ipc-shm to PNWHITELIST
Issue: LINCD-5367

The ipc-shm package is not in PNWHITELIST by default,
such that it may be considered obsoleted and will be droped.
So add it to PNWHITELIST, and if users want to use it, just
adding one line in <build-dir>/conf/local.conf:

  IMAGE_INSTALL_append = ' ipc-shm'

then the package will be installed into WIC image.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>