nxp-s32g2xx: README: update to compatible with pre-released SDK bsp29
Issue: LINUXEXEC-13906

Due to the pfe and llce are upgraded to new version, so update
readme file to compatible with these changes. Moreover, for the
layout of memory has been changed because of the updates of linux
kernel and other features, the kernel load address and dtb load
address need to update too.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>