nxp-s32g2xx: pkcs11-hse: update to bsp31 rc3
Issue: LINUXEXEC-17089

The NXP SDK bsp31 rc4 has been released, but there are some issues in rc4,
which will make the pkcs11-hse application build fail and the related
libpkcs-hse.so include some undefined symbols, so this patch will just update
pkcs11-hse recipe to compatible with NXP SDK bsp31 rc3 version, and we will
update pkcs recipe to new version when the above stated issues fixed.

And in this update, the pkcs demo has been restructured into two libs:
	libpkcs-hse.so and libhse.so
which will interact with the linux uio-hse driver to support the hse demo.
More over, due to the rc3 update, the bb file is modified accordingly and
add a patch:
	0001-pkcs11-hse-Makefile-using-internal-compile-variables.patch
to use bitbake internal variable CC to replace the CROSS_COMPILE variable
in Makefile of source codes, or else the software will build fail with the
CROSS_COMPILE not set.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>